requirements.txt
ccxt
asyncio
python-dotenv